# feat(search): Added App list to metadata endpoint #192