--- a/ozpcenter/api/storefront/tests/test_api_storefront.py
+++ b/ozpcenter/api/storefront/tests/test_api_storefront.py
@@ -37,6 +37,7 @@ class StorefrontApiTest(APITestCase):
         self.assertIn('contact_types', response.data)
         self.assertIn('listing_types', response.data)
         self.assertIn('intents', response.data)
+        self.assertIn('listing_titles', response.data)
         for i in response.data['agencies']:
             self.assertTrue('listing_count' in i)