@rpokorny 2016-07-20
Did you guys consider the performance ramifications of this?