@wski 2016-07-20
![screenshot from 2016-07-20 15-23-32](https://cloud.githubusercontent.com/assets/1770198/16999842/00a73acc-4e8e-11e6-9194-f6e306822e8e.png)

Frontend updated to support feat.