@mannyrivera2010 2016-07-21
### Performance Snippets:
**Typeahead:**
````html
<!DOCTYPE html>
<html>
   <head>
      <script src="https://code.jquery.com/jquery-2.2.4.js"   integrity="sha256-iT6Q9iMJYuQiMWNd9lDyBUStIq/8PuOW33aOqmvFpqI="   crossorigin="anonymous"></script>
      <script type="text/javascript" src="http://twitter.github.io/typeahead.js/releases/latest/typeahead.jquery.js"></script>
      <script type="text/javascript">
         var substringMatcher = function(strs) {
           return function findMatches(q, cb) {
             var matches, substringRegex;

             // an array that will be populated with substring matches
             matches = [];

             // regex used to determine if a string contains the substring `q`
             substrRegex = new RegExp(q, 'i');

             // iterate through the pool of strings and for any string that
             // contains the substring `q`, add it to the `matches` array
             $.each(strs, function(i, str) {
               if (substrRegex.test(str)) {
                 matches.push(str);
               }
             });

             cb(matches);
           };
         };

         var states = [
                  "Air Mail",
                  "Bread Basket",
                  "Chart Course",
                  "Chatter Box",
                  "Clipboard",
                  "FrameIt",
                  "Hatch Latch",
                  "JotSpot",
                  "LocationLister",
                  "LocationViewer",
                  "LocationAnalyzer",
                ];

        var statesOut = []

        $.each(states, function( index, value ) {
          for (i = 0; i <= 500; i++) {
                statesOut.push(value+' '+i);
          }
        });

        console.log(statesOut.length)

      </script>
   </head>
   <body>
      <div id="the-basics">
         <input class="typeahead" type="text" placeholder="Type App Name">
      </div>

      <script type="text/javascript">
          $('#the-basics .typeahead').typeahead({
            hint: true,
            highlight: true,
            minLength: 1
          },
          {
            name: 'apps',
            source: substringMatcher(statesOut)
          });
      </script>
   </body>
</html>
````
TypeAhead has good performance for 5541 entries.
Normally we have less than 200 entries that a user can see

**Metadata endpoint** 
this endpoint will be called on user's first time on the page and on page refreshes. 

/api/metadata/ takes around 200 to 300 ms for a response after adding listing titles for the user without cache, and takes around 160 to 230 ms for a response before adding listing titles for the user