@mannyrivera2010 2016-07-21
We could also use tips from https://www.dabapps.com/blog/api-performance-profiling-django-rest-framework/ to improve performance