@rpokorny 2016-07-21
I was mainly worried about network bandwidth.  I suppose if its only 200 listings it won't be _too_ big a problem.  I was assuming you needed to worry about it being thousands.